Reorganize module structure
Now CLI definition is separated from command implementation
