+++ dsync-cli/src/cli/command/file.rs
@@ -1,7 +1,8 @@
 use clap::Subcommand;
 
-use super::{GroupId, RemoteId};
 use std::path::PathBuf;
+
+use crate::command::model::{GroupId, RemoteId};
 
 #[derive(Subcommand)]
 pub(crate) enum FileCommand {
