+++ dsync-cli/src/cli/command/host.rs
