+++ dsync-cli/src/command/file.rs
