+++ dsync-cli/src/command/host.rs
