+++ dsync-cli/src/command/model.rs
@@ -1,0 +1,2 @@
+pub(crate) type GroupId = String;
+pub(crate) type RemoteId = String;
